Update to v104r08 release.
byuu says:

Changelog:

  - processor/upd96050: code cleanups
  - processor/upd96050: improved emulation of S1/OV1 flags [thanks to
    Cydrak, Lord Nightmare]
  - tomoko/settings/audio: reduced the size of the frequency/latency
    combo boxes to show longer device driver names

Errata: I need to clear regs.sr in uPD96050::power()

Note: the S1/OV1 emulation is likely not 100% correct yet, but it's a
step in the right direction. No SNES games actually use S1/OV1, so this
shouldn't result in any issues, I'd just like to have this part of the
chip emulated correctly.

+++ higan/target-tomoko/settings/audio.cpp
@@ -57,7 +57,7 @@ AudioSettings::AudioSettings(TabFrame* parent) : TabFrameItem(parent) {
 auto AudioSettings::updateDevice() -> void {
   frequencyList.reset();
   for(auto& frequency : audio->availableFrequencies()) {
-    frequencyList.append(ComboButtonItem().setText(frequency));
+    frequencyList.append(ComboButtonItem().setText((uint)frequency));
     if(frequency == settings["Audio/Frequency"].real()) {
       frequencyList.item(frequencyList.itemCount() - 1).setSelected();
     }
